Move on to next file if any errors are encountered
If unable to process the current file for whatever reason, move on to the
next one.

Previously if we were unable to download a file, we would still try to
uncompress it.

This helps work around a rare race condition that exists if multiple processes
are working on the same S3 bucket. One process downloads, processes the file, and
deletes the S3 object. The other process has already received the list of S3 files
to work on but hasn't downloaded the file yet. When it tries to, file no longer
exists and we see a 404 error. Since we didn't break out of loop, we then try
to uncompress the file that couldn't be downloaded.

Closes https://github.com/zumba/cloudfront-loggly-uploader/issues/1

Signed-off-by: Rene Fragoso <ctrlrsf@gmail.com>